server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
instaloader==4.14.1
python-jose[cryptography]==3.4.0
pydantic==2.11.3
aiofiles==24.1.0
aiohttp==3.11.18
